+++ padme/tests.py
@@ -15,6 +15,13 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with Padme.  If not, see <http://www.gnu.org/licenses/>.
+"""
+:mod:`padme.tests` -- tests for padme
+=====================================
+
+This package contains unit tests for padme. You don't want to use it directly
+and it can get changed at any time without notice.
+"""
 from __future__ import print_function, absolute_import, unicode_literals
 
 import doctest
@@ -134,6 +141,17 @@ class proxy_as_function(unittest.TestCase):
         self.assertGreaterEqual(self.proxy, -1)
         self.assertGreaterEqual(self.obj, 0)
         self.assertGreaterEqual(self.obj, -1)
+
+    @unittest.skipUnless(sys.version_info[0] == 2, "requires python 2")
+    def test_cmp(self):
+        class C(object):
+            def __cmp__(self, other):
+                return -1
+        self.obj = C()
+        self.proxy = proxy(self.obj)
+        other = object()
+        self.assertEqual(cmp(self.proxy, other), cmp(self.obj, other))
+        self.assertEqual(self.proxy.__cmp__(other), cmp(self.obj, other))
 
     def test_hash(self):
         self.assertEqual(hash(self.proxy), hash(self.obj))
@@ -320,6 +338,7 @@ class proxy_as_function(unittest.TestCase):
             __ne__
             __gt__
             __ge__
+            __cmp__
             __hash__
             __bool__
             __nonzero__