+++ HISTORY.rst
@@ -3,6 +3,20 @@
 History
 =======
 
+1.1 (YYYY-MM-DD)
+----------------
+* Add general support for **Python 2.7**.
+* Make ``__nonzero__()`` and ``__unicode__()`` exlusive to Python 2.x.
+* Make ``__bool__()`` and ``__bytes__()`` exclusive to Python 3.x.
+* Make ``__length_hint()`` exclusive to Python 3.4.
+* Add support for the ``__cmp__()`` method, exclusive to Python 2.x.
+* Add support for accessing the proxied object with the new
+  :meth:`~padme.proxy.original()` function.
+* Add support for accessing proxy state with the new
+  :meth:`~padme.proxy.state()` function.
+* De-couple proxy classes from proxied objects, much more lightweight proxy
+  design is possible this way (less objects, lower cost to create each new proxy).
+
 1.0 (2014-02-11)
 ----------------